Kernel argument list manipulation stored in a separate module
