+++ include/kenv.h
@@ -1,0 +1,10 @@
+
+
+void setup_kenv(int pfm_argc, char **pfm_argv);
+
+char *kenv_get(const char *key);
+
+int kenv_get_argc(void);
+char **kenv_get_argv(void);
+char **kenv_get_user_argv(void);
+char **kenv_get_user_envv(void);
