+++ sys/Makefile
@@ -25,6 +25,7 @@ SOURCES = \
 	initrd.c \
 	interrupt.c \
 	kbss.c \
+	kenv.c \
 	klog.c \
 	ktest.c \
 	main.c \
